Move Config files into ConfigForm folder
